+ fixed several warnings with highest warn-level + fixed warnings in PyCXX

+++ src/Base/Reader.cpp
@@ -116,7 +116,7 @@ const char* Base::XMLReader::localName(void)
 
 unsigned int Base::XMLReader::getAttributeCount(void) const
 {
-  return AttrMap.size();
+  return (unsigned int)AttrMap.size();
 }
 
 long Base::XMLReader::getAttributeAsInteger(const char* AttrName) const
